fix time duration convert for ping

+++ inputs/ping/ping.go
@@ -59,7 +59,7 @@ func (ins *Instance) Init() error {
 	if ins.Timeout == 0 {
 		ins.calcTimeout = time.Duration(3) * time.Second
 	} else {
-		ins.calcTimeout = time.Duration(ins.Timeout) * time.Second
+		ins.calcTimeout = time.Duration(ins.Timeout * float64(time.Second))
 	}
 
 	if ins.Interface != "" {
